docs: Update backlog
Removes the completed feature for multiple output formats from the backlog.
Also removes the result schema feature, as it was partially implemented to support the new formats.
